removed components to make the posting more clear
